requirements.txt
typer==0.12.3
rich==13.7.1
textual==0.76.0
Pillow==10.4.0
numpy==2.1.2
faiss-cpu==1.9.0.post1
open_clip_torch==2.24.0
torch>=2.2.0
torchvision>=0.17.0
imagehash==4.3.1
exifread==3.0.0
jinja2==3.1.4
tqdm==4.66.5
# Optional (heavy): enable face clustering by installing these manually
# face-recognition
# dlib
